requirements.txt
Django==3.2.16
pytest-django==4.5.2
pytest-pythonpath==0.7.3
